Splitting PWGGAGammaConv library. New library PWGGAGammaConvBase to avoid circular dependencies for nano AODs
