+++ PWGGA/CMakeLists.txt
@@ -16,6 +16,7 @@
 add_subdirectory(CaloTrackCorrelations)
 add_subdirectory(EMCALTasks)
 add_subdirectory(GammaConv)
+add_subdirectory(GammaConvBase)
 add_subdirectory(Hyperon)
 add_subdirectory(PHOSTasks)
 add_subdirectory(PWGGAUtils)
