+++ PWGGA/GammaConvBase/AliCaloPhotonCuts.cxx
