+++ PWGGA/GammaConvBase/AliCaloPhotonCuts.h
