+++ PWGGA/GammaConvBase/AliConvEventCuts.cxx
