+++ PWGGA/GammaConvBase/AliConversionPhotonBase.cxx
