+++ PWGGA/GammaConvBase/AliConversionPhotonCuts.cxx
