+++ PWGGA/GammaConvBase/AliConversionPhotonCuts.h
