+++ PWGGA/GammaConvBase/AliDalitzElectronCuts.cxx
