+++ PWGGA/GammaConvBase/AliDalitzElectronCuts.h
