+++ PWGGA/GammaConvBase/AliDalitzElectronSelector.cxx
