+++ PWGGA/GammaConvBase/AliPhotonIsolation.cxx
