+++ PWGGA/GammaConvBase/PWGGAGammaConvBaseLinkDef.h
@@ -1,0 +1,35 @@
+#ifdef __CINT__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+// Base classes
+#pragma link C++ class AliConversionPhotonBase+;
+#pragma link C++ class AliAODConversionParticle+;
+#pragma link C++ class AliAODConversionMother+;
+#pragma link C++ class AliAODConversionPhoton+;
+#pragma link C++ class AliKFConversionPhoton+;
+#pragma link C++ class AliKFConversionMother+;
+#pragma link C++ class AliCaloPhotonCuts+;
+#pragma link C++ class AliConvEventCuts+;
+#pragma link C++ class AliConversionPhotonCuts+;
+#pragma link C++ class AliConversionCuts+;
+#pragma link C++ class AliConversionSelection+;
+#pragma link C++ class AliV0ReaderV1+;
+#pragma link C++ class AliConversionAODBGHandlerRP+;
+#pragma link C++ class AliConversionTrackCuts+;
+#pragma link C++ class AliConversionMesonCuts+;
+#pragma link C++ class AliDalitzElectronCuts+;
+#pragma link C++ class AliDalitzElectronSelector+;
+#pragma link C++ class AliCaloTrackMatcher+;
+#pragma link C++ class AliPhotonIsolation+;
+
+// User tasks
+#pragma link C++ class AliDalitzAODESD+;
+#pragma link C++ class AliDalitzAODESDMC+;
+#pragma link C++ class AliDalitzEventMC+;
+#pragma link C++ class AliDalitzData+;
+
+
+#endif
